+++ glfwtest/graphics/models/SubMesh.h
@@ -13,7 +13,7 @@ public:
 	SubMesh(uint32_t triangleCount);
 	~SubMesh();
 
-	bool AllocateBuffers(void* vertexData, const size_t& vertexDataSize, void* indexData, const size_t& indexDataSize);	
+	bool AllocateBuffers(void* vertexData, const uint32_t& vertexDataSize, void* indexData, const uint32_t& indexDataSize);
 	bool IsLoaded() const { return vertexBuffer != nullptr && indexBuffer != nullptr; }
 
 	void Draw(std::shared_ptr<CommandBuffer> commandBuffer);
